+++ src/cadenza.test.ts
@@ -64,7 +64,7 @@ describe('Given a Cadenza JS client instance', () => {
 
   it('Throws when attempting to show an embedding target without an iframe', () =>
     expect(() => cadenza(BASE_URL).show(EMBEDDING_TARGET_ID)).toThrow(
-      'present',
+      'Required iframe',
     ));
 
   it('Throws when attempting to show an embedding target in an invisible iframe', () => {
@@ -212,6 +212,16 @@ describe('Given a Cadenza JS client instance', () => {
     describe('When the event is sent again after unsubscribing', () => {
       beforeEach(() => {
         unsubscribe();
+        sendEvent(EVENT.type, EVENT.detail);
+      });
+
+      it('Subscribers are not called anymore', () =>
+        expect(subscriber).not.toHaveBeenCalled());
+    });
+
+    describe('When the event is sent again after the instance has been destroyed', () => {
+      beforeEach(() => {
+        cad.destroy();
         sendEvent(EVENT.type, EVENT.detail);
       });